Add components
New components added:
- airflow_control_select
- air_purifier_switch
- night_mode_switch
- circulator_switch

Thank you to @duhow for discovering the command `0x42`.

This commit contains code that is tested on a `MSZ-LN25VG2W`. It mostly reuses old code, but also introduces some new stuff.
I am by no means a coder and the commited code should probably be tidied up by someone who knows what they are doing.

airflow_control_select:
Enables `Airflow control mode`. Known settings are `INDIRECT`, `DIRECT` and `EVEN`.
The horizontal vane must be in a special mode (`0x80`) for the airflow modes to be active - here dubbed `AIRFLOW CONTROL`.
Operating instructions/manual will list `OFF` as a mode, but this is actually just the unit being put in another horizontal vane mode - usually `SWING`.
The unit will remember what mode it was in when horizontal vane mode is changed, but the byte that holds the state will allways be `0x00` when not in `AIRFLOW CONTROL`.

air_purifier_switch:
Adds a switch that enable/disable the air purifier.

night_mode_switch:
Adds a switch that enable/disable nigth mode.
Documented to work: dimming of the led indicators on the indoor unit.
The performance of the unit should also change, but this has not been verified.
Using the remote control will override the setting!

circulator_switch.
Adds a switch that enable/disable the circulator mode.
This mode is being controlled by a button on the remote that usually controls `ECONOCOOL`.
It might be the same bit that is being set and just the units that function differently.
Needs to be tested/documented.
Using the remote of the `MSZ-LNxxVG2W`, this mode can only be toggled when in `HEAT` mode.
But, we can actually set it in other modes.